Create BucketSort.cpp
Creating Buckets: We create a number of buckets based on the maximum value in the array. Each bucket is a vector that will hold integers.
Distributing Elements: We determine the bucket for each number by dividing the number by 10 (the bucket size).
Sorting Buckets: Each bucket is sorted using std::sort.
Combining: Finally, we combine all the buckets back into the original array.