Added Activity selection Algorithm in java
Detailed Explanation:
Initialization:

count is initialized to 0, but immediately set to 1 after selecting the first activity by default.
ArrayList<Integer> al is created to store the indices of the selected activities. This list can be useful if you want to keep track of the actual activities selected.
First Activity Selection:

The first activity is always selected because it's the one with the earliest end time (as per the assumption that e[] is sorted).
count = 1 because the first activity is selected.
al.add(0) adds the index of the first activity to the list of selected activities.
lastele = e[0] keeps track of the end time of the first selected activity.
Loop Through Remaining Activities:

The loop starts from the second activity (i = 1) and goes through all activities.
For each activity, it checks if the start time (s[i]) is greater than or equal to the end time of the last selected activity (lastele). This ensures no overlap between the selected activities.
If the current activity can be selected, the following steps are performed:
count++ increases the number of selected activities.
al.add(i) adds the index of the current activity to the list of selected activities.
lastele = e[i] updates the end time of the last selected activity to the end time of the current activity.
Return the Count:

The function returns the value of count, which represents the maximum number of activities that can be performed by a single person without overlapping.